app/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.30
python-jose[cryptography]==3.3.0
python-dotenv>=1.0.1
anthropic>=0.49.0
pydantic[email]>=2.10.0
python-multipart>=0.0.9
bcrypt>=4.0.0
